Created a define in includes.h for the maximum name length for creatures. Used this value in the declaration of creaturest.
Created a function called enter_name which allows the player to 
enter a name for a new squad member. Gives the option of a default value
if no name is entered.

Used enter_name for dates and kidnapping. Now the player can enter
a name without the possibility of a buffer overflow and that name
will default to the person's real name (propername) if no name
is given. (I like this because I am lazy and don't feel like making
creative names for people I am just going to send to get killed anyway)

Additionally, a new string is being written in all cases that a name will be
entered. This string says that if the player just presses enter, 
the real name for the new Liberal will be used.


git-svn-id: http://svn.code.sf.net/p/lcsgame/code/trunk@57 a218ab89-6f2a-4548-b3d1-e91a371d7644