Added and changed more help options to help.cpp.
I've made a little help tool to create a lot of text that looks and feels like LCS that generates c++ for LCS in real time. If you want it PM myself (firewolf) or jonathan on the forum, the code is freely available, just request any features from myself if you want any.

git-svn-id: http://svn.code.sf.net/p/lcsgame/code/trunk@397 a218ab89-6f2a-4548-b3d1-e91a371d7644